New forum refactor part 1: old code, new architecture
The whole codebase needs to be rewritten for the new forums, which presents an
opportunity to clear up some of the bad design decisions I made at the start.

This commit uses the newly-written BabelExt architecture, but is still written
to work on the old forums.

Functionality due for removal has been dropped now to make porting easier.