Rewrite "notification" (now "user action") section for usability and bans
The previous design made it too easily to accidentally issue an infraction instead of a warning,
and had no facility to ban users when giving them an infraction.

This commit also pushes the user more strongly to log in to ModCP and keeps them logged in
(so bans can happen), gets up to 50 recent warnings/infractions and fixes an
issue with ModCP pages changing ID values.  These could have been split into
multiple, commits, but it was easier this way and I don't expect it would have
actually helped anything.